Enemy Laser Color Change & Targeting System
Changed Enemy Laser Colors to further distinguish them from the Player.
Also added the ground works for a Targeting System. The Targeting System
is still a work in progress. The idea is for an additional laser to be
shot from the center of the ship. This laser will make it's way to the
closest Enemy, with no need of aiming. This feature will not be
available until completed.